requirements.txt
Django
Pillow
django-filter
django-import-export
djangorestframework
gunicorn